+++ scripts/BruteForce/global.lua
@@ -31,6 +31,9 @@ end
 local function addBounty(data)
     AddBounty(data.player, data.bounty)
 end
+local function triggerTrap(data)
+    TriggerTrap(data.o, data.player)
+end
 
 I.Activation.addHandlerForType(types.Door, lockableOpen)
 I.Activation.addHandlerForType(types.Container, lockableOpen)
@@ -41,7 +44,8 @@ return {
         onSave = onSave,
     },
     eventHandlers = {
-        checkJammedLock = checkJammedLock,
-        addBounty = addBounty,
+        CheckJammedLock = checkJammedLock,
+        AddBounty = addBounty,
+        TriggerTrap = triggerTrap,
     },
 }